+++ src/App/App.js
@@ -2,16 +2,16 @@ import React from 'react';
 
 import { Provider } from 'react-redux';
 
-import Home from './components/Home'
-import AddNews from './components/AddNews';
+import Home from '../components/Home'
+import AddNews from '../components/AddNews';
 import {
   BrowserRouter as Router,
   Switch,
   Route
 } from "react-router-dom";
-import Appbar from './components/Appbar'
-import {getAllRoutes} from './Routes'
-import store from './store'
+import Appbar from '../components/Appbar'
+import {getAllRoutes} from '../navigation/Routes'
+import store from '../store/store'
 
 function App() {
   return (
